requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
certifi==2025.6.15
charset-normalizer==3.4.2
circuitbreaker==2.1.3
edgartools==4.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.12.3
idna==3.10
Jinja2==3.1.6
lxml==6.0.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
nest-asyncio==1.6.0
numpy==2.3.1
orjson==3.10.18
packaging==25.0
pandas==2.3.0
psycopg2-binary==2.9.10
pyarrow==20.0.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
python-dateutil==2.9.0.post0
pytz==2025.2
rank-bm25==0.2.2
RapidFuzz==3.13.0
ratelimit==2.2.1
requests==2.32.4
rich==14.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
stamina==25.1.0
tabulate==0.9.0
tenacity==9.1.2
textdistance==4.6.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
Unidecode==1.4.0
urllib3==2.5.0
watchdog==6.0.0
